[UPD] - Update variable types from Int to Long and adjust default values across multiple Kotlin files to enhance code clarity and maintain compliance with Explicit API mode.

+++ android/app/src/main/kotlin/com/jabook/app/jabook/audio/CrossFadePlayer.kt
@@ -45,7 +45,7 @@ public class CrossFadePlayer(
     private var currentPlayer: ExoPlayer = playerA
     private var nextPlayer: ExoPlayer = playerB
 
-    public var crossFadeDurationMs: Int = 0
+    public var crossFadeDurationMs: Long = 0L
     private var currentAnimator: ValueAnimator? = null
     private var isCrossFading = false
